Renamed all manifests and source code to gnosis/stories
