+++ README.md
@@ -7,11 +7,14 @@ developer who works with or wants to learn to work with Kubernetes.
 ## Table of Contents
 * [kubernetes 101](./kubernetes-101) - a collection of howtos, tips, and
   tutorials for
-  * [developers](./kubernetes-101/developers)
-  * [operators](./kubernetes-101/operators)
-* [Contributing to Kubernetes](./contributing-to-kubernetes/) - a collection of
-  common questions about the Kubernetes community. This is intended to be a
-  roadmap for the existing documentation.
+  * [developers](./kubernetes-101/developers) people who are building things on
+    top off Kubernetes.
+  * [operators](./kubernetes-101/operators) people who are running Kubernetes
+    and taking care of it.
+* [Stories](./stories/) - a collection of developer stories about Kubernetes.
+  Here you will find stories detailing how parts of Kubernetes work. These are
+  inspired by real-life issues, bugs, and things that are not very well know.
+  These are written for people who want to begin contributing to Kubernetes.
 * [go-client](./go-client/) - a collection of handy programs for building
   things on top of Kubernetes using the Go client.
 * [kind](./kind/) - [github.com/kubernetes-sigs/kind] relates resources,
