--- contributing-to-kubernetes/e2e-tests/example-e2e-test/go.mod
@@ -1,10 +1,0 @@
-module github.com/contributing-to-kubernetes/cool-kubernetes/contributing-to-kubernetes/e2e-tests/example-e2e-test
-
-go 1.13
-
-require (
-	github.com/onsi/ginkgo v1.12.0
-	github.com/onsi/gomega v1.9.0
-	k8s.io/apimachinery v0.17.4
-	k8s.io/klog v1.0.0
-)
