--- contributing-to-kubernetes/kk-pr-85968/example-cobra/go.mod
@@ -1,12 +1,0 @@
-module github.com/contributing-to-kubernetes/cool-kubernetes/contributing-to-kubernetes/kk-pr-85968/example-cobra
-
-go 1.13
-
-require (
-	github.com/spf13/cobra v0.0.6
-	github.com/spf13/pflag v1.0.5
-	k8s.io/apiserver v0.17.4
-	k8s.io/component-base v0.17.4
-	k8s.io/klog v1.0.0
-	k8s.io/utils v0.0.0-20191114184206-e782cd3c129f
-)
