+++ go-client/dynamic-client/Makefile
@@ -1,6 +1,6 @@
 GIT_BRANCH := $(shell git rev-parse --abbrev-ref HEAD 2>/dev/null)
 GIT_BRANCH_CLEAN := $(shell echo $(GIT_BRANCH) | sed -e "s/[^[:alnum:]]/-/g")
-DOCKER_IMAGE := cool-kubernetes$(if $(GIT_BRANCH_CLEAN),:$(GIT_BRANCH_CLEAN))
+DOCKER_IMAGE := cool-kube$(if $(GIT_BRANCH_CLEAN),:$(GIT_BRANCH_CLEAN))
 
 DOCKERFILE ?= Dockerfile
 DOCKER_BUILD_ARGS ?= --force-rm -f "$(DOCKERFILE)"
@@ -17,5 +17,5 @@ build:
 
 .PHONY: push
 push: build
-	docker tag $(DOCKER_IMAGE) alejandrox1/cool-kubernetes:v1
-	docker push alejandrox1/cool-kubernetes:v1
+	docker tag $(DOCKER_IMAGE) alejandrox1/cool-kube:v1
+	docker push alejandrox1/cool-kube:v1
