+++ go-client/dynamic-client/go.mod
@@ -1,9 +1,9 @@
-module github.com/alejandrox1/cool-kubernetes/controller-runtime
+module github.com/contributing-to-kubernetes/gnosis/go-client/dynamic-client
 
-go 1.12
+go 1.13
 
 require (
-	k8s.io/api v0.0.0-20190918155943-95b840bb6a1f
-	k8s.io/apimachinery v0.0.0-20190913080033-27d36303b655
-	k8s.io/client-go v0.0.0-20190918160344-1fbdaa4c8d90
+	k8s.io/api v0.17.4
+	k8s.io/apimachinery v0.17.4
+	k8s.io/client-go v0.17.4
 )
