+++ go-client/dynamic-client/skaffold.yaml
@@ -2,7 +2,7 @@ apiVersion: skaffold/v1beta14
 kind: Config
 build:
   artifacts:
-  - image: alejandrox1/cool-kubernetes
+  - image: alejandrox1/cool-kube
 deploy:
   kubectl:
     manifests:
