+++ go-client/informer/go.mod
@@ -1,13 +1,9 @@
-module github.com/alejandrox1/cool-kubernetes/go-client/informer
+module github.com/contributing-to-kubernetes/gnosis/go-client/informer
 
-go 1.12
+go 1.13
 
 require (
-	golang.org/x/crypto v0.0.0-20190911031432-227b76d455e7 // indirect
-	golang.org/x/oauth2 v0.0.0-20190604053449-0f29369cfe45 // indirect
-	golang.org/x/time v0.0.0-20190308202827-9d24e82272b4 // indirect
-	k8s.io/api v0.0.0-20190905160310-fb749d2f1064
-	k8s.io/apimachinery v0.0.0-20190831074630-461753078381
-	k8s.io/client-go v0.0.0-20190620085101-78d2af792bab
-	k8s.io/utils v0.0.0-20190907131718-3d4f5b7dea0b // indirect
+	k8s.io/api v0.17.4
+	k8s.io/apimachinery v0.17.4
+	k8s.io/client-go v0.17.4
 )
