+++ stories/e2e-tests/example-e2e-test/e2e_test.go
@@ -25,10 +25,10 @@ import (
 
 	"k8s.io/klog"
 
-	"github.com/contributing-to-kubernetes/cool-kubernetes/contributing-to-kubernetes/e2e-tests/example-e2e-test/framework"
+	"github.com/contributing-to-kubernetes/gnosis/stories/e2e-tests/example-e2e-test/framework"
 
 	// Test Sources.
-	_ "github.com/contributing-to-kubernetes/cool-kubernetes/contributing-to-kubernetes/e2e-tests/example-e2e-test/apps"
+	_ "github.com/contributing-to-kubernetes/gnosis/stories/e2e-tests/example-e2e-test/apps"
 )
 
 // Flags is the flag set that AddOptions adds to. Test authors should
