+++ stories/e2e-tests/example-e2e-test/go.mod
@@ -1,0 +1,11 @@
+module github.com/contributing-to-kubernetes/gnosis/stories/e2e-tests/example-e2e-test
+
+go 1.13
+
+require (
+	github.com/onsi/ginkgo v1.10.1
+	github.com/onsi/gomega v1.7.0
+	k8s.io/apimachinery v0.17.4
+	k8s.io/client-go v0.17.4 // indirect
+	k8s.io/klog v1.0.0
+)
