+++ stories/kk-pr-85968/example-cobra/go.mod
@@ -1,0 +1,12 @@
+module github.com/contributing-to-kubernetes/gnosis/stories/kk-pr-85968/example-cobra
+
+go 1.13
+
+require (
+	github.com/spf13/cobra v0.0.7
+	github.com/spf13/pflag v1.0.5
+	k8s.io/apiserver v0.18.1
+	k8s.io/component-base v0.18.1
+	k8s.io/klog v1.0.0
+	k8s.io/utils v0.0.0-20200324210504-a9aa75ae1b89
+)
